requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.14.0
django-crispy-forms==2.0
djangorestframework==3.14.0
fake-useragent==1.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
pendulum==2.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pytzdata==2020.1
PyVirtualDisplay==3.0
PyYAML==6.0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wsproto==1.2.0
